AYS-147 | Assignment, RequestBodyAssignments and Pagination class updated.
