+++ src/test/java/org/ays/payload/Pagination.java
@@ -10,4 +10,18 @@ public class Pagination {
     private int page;
     private int pageSize;
 
+    public static Pagination createPagination() {
+        Pagination pagination = new Pagination();
+        pagination.setPage(1);
+        pagination.setPageSize(10);
+        return pagination;
+    }
+
+    public static Pagination setPagination(int page, int pageSize) {
+        Pagination pagination = new Pagination();
+        pagination.setPageSize(pageSize);
+        pagination.setPage(page);
+        return pagination;
+    }
+
 }
