+++ src/test/java/org/ays/payload/RequestBodyAssignments.java
@@ -10,4 +10,23 @@ public class RequestBodyAssignments {
     private Pagination pagination;
     private FiltersForAssignments filter;
 
+    public static RequestBodyAssignments generateRequestBodyAssignments(int page, int pageSize) {
+        RequestBodyAssignments requestBodyAssignments = new RequestBodyAssignments();
+        Pagination pagination = new Pagination();
+        pagination.setPage(page);
+        pagination.setPageSize(pageSize);
+        requestBodyAssignments.setPagination(pagination);
+        return requestBodyAssignments;
+    }
+
+    public static RequestBodyAssignments generateRequestBodyAssignmentsWithPhoneNumberFilter(PhoneNumber phoneNumber) {
+        RequestBodyAssignments requestBodyAssignments = new RequestBodyAssignments();
+
+        requestBodyAssignments.setPagination(Pagination.createPagination());
+        FiltersForAssignments filters = new FiltersForAssignments();
+        filters.setPhoneNumber(phoneNumber);
+        requestBodyAssignments.setFilter(filters);
+        return requestBodyAssignments;
+    }
+
 }
