+++ src/test/java/org/ays/tests/institution/assignmentmanagementservice/DeleteAssignmentTest.java
@@ -18,7 +18,7 @@ public class DeleteAssignmentTest {
 
     @BeforeMethod(alwaysRun = true)
     public void setup() {
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
         assignmentId = Helper.extractAssignmentIdByPhoneNumber(assignment.getPhoneNumber());
     }
