+++ src/test/java/org/ays/tests/institution/assignmentmanagementservice/PutAssignmentTest.java
@@ -16,7 +16,7 @@ public class PutAssignmentTest {
 
     @BeforeMethod(alwaysRun = true)
     public void setup() {
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
         assignmentId = Helper.extractAssignmentIdByPhoneNumber(assignment.getPhoneNumber());
     }
