+++ src/test/java/org/ays/tests/user/assignmentmanagementservice/PostAssignmentApproveTest.java
@@ -24,7 +24,7 @@ public class PostAssignmentApproveTest {
     public void setup() {
         userCredentials = UserCredentials.generateCreate();
         location = new Location();
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
 
     }
