+++ src/test/java/org/ays/tests/user/assignmentmanagementservice/PostAssignmentCompleteTest.java
@@ -25,7 +25,7 @@ public class PostAssignmentCompleteTest {
     public void setup() {
         userCredentials = UserCredentials.generateCreate();
         location = new Location();
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
 
     }
