+++ src/test/java/org/ays/tests/user/assignmentmanagementservice/PostAssignmentStartTest.java
@@ -25,7 +25,7 @@ public class PostAssignmentStartTest {
     public void setup() {
         userCredentials = UserCredentials.generateCreate();
         location = new Location();
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
 
     }
