+++ src/test/java/org/ays/tests/user/locationmanagementservice/PostUserLocationTest.java
@@ -25,7 +25,7 @@ public class PostUserLocationTest {
     public void setup() {
         userCredentials = UserCredentials.generateCreate();
         location = new Location();
-        assignment = Helper.createANewAssignment();
+        assignment = Assignment.generateCreateAssignment();
 
     }
